Feat: Reduce unnecessary requests for wiki pages.
