+++ web/src/pages/dataset/compilation/constants.ts
@@ -1,5 +1,4 @@
 export enum ViewMode {
-  Graph = 'graph',
   LlmWiki = 'llm-wiki',
   Skills = 'skills',
 }
